lab4/src/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
pip==21.2.4
psycopg2-binary==2.9.2
pyTelegramBotAPI==4.2.0
requests==2.26.0
setuptools==58.0.4
urllib3==1.26.7
wheel==0.37.0
